XML file.
However, the provided diff actually shows no difference. No lines were added or deleted. Both the lines before and after are identical. The commit message therefore needs to be revised:

Commit message:
No changes in German localization of DreadOverlord mod

The diff shows no changes were made to the DreadOverlord mod's German localization. The content of the XML file remains the same as it was in the last revision. No changes in translations or formatting have been made to any of the entries in this file. The commit message has been mistaken in mentioning changes that are not present in the diff.