Add .gitignore rule and further German translations
This commit introduces a new .gitignore rule that ignores any file ending with .loca. Additionally, a bulk of the features and various content in German language was translated as part of the MoreFeats.xml file. The purpose here is to improve the overall experience of German speaking users and contributors by adding localized content and clean contribution workflow.

While most of the changes are translations, the commit also introduced a new feature called Dwarven Fortitude - "Zwergische Tapferkeit". This suggests a new feature added to the gameplay, again to improve the gaming experience showcasing diversity in cultural environments displayed in the game.