Add new localization data for the BasketEquipmentNSFWGER mod
A large amount of new localization data has been added to the BasketEquipmentNSFWGER mod. The changes consist of descriptions for various pieces of equipment, such as clothing and weaponry, in English. These descriptions involve information about the item's appearance and function, as well as the type of character that can equip it. The update allows English-speaking players to understand the equipment options more clearly when using the BasketEquipmentNSFWGER mod.